+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/rating/commands/DisableRating.kt
@@ -1,13 +1,18 @@
 package com.github.insanusmokrassar.AutoPostTelegramBot.plugins.rating.commands
 
+import com.github.insanusmokrassar.AutoPostTelegramBot.base.database.exceptions.NoRowFoundException
+import com.github.insanusmokrassar.AutoPostTelegramBot.base.database.tables.PostsTable
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.commonLogger
 import com.github.insanusmokrassar.AutoPostTelegramBot.plugins.rating.database.PostsLikesMessagesTable
 import com.github.insanusmokrassar.AutoPostTelegramBot.plugins.rating.disableLikesForPost
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.commands.Command
 import com.github.insanusmokrassar.TelegramBotAPI.bot.RequestsExecutor
+import com.github.insanusmokrassar.TelegramBotAPI.requests.DeleteMessage
+import com.github.insanusmokrassar.TelegramBotAPI.requests.send.SendMessage
 import com.github.insanusmokrassar.TelegramBotAPI.types.UpdateIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.AbleToReplyMessage
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.CommonMessage
+import com.github.insanusmokrassar.TelegramBotAPI.utils.extensions.executeUnsafe
 import java.lang.ref.WeakReference
 
 class DisableRating(
@@ -21,16 +26,35 @@ class DisableRating(
         val executor = executorWR.get() ?: return
         val replied = (message as? AbleToReplyMessage) ?.replyTo ?: return // TODO:: Add sending tooltip
 
-        val postId = postsLikesMessagesTable.postIdByMessageId(
-            replied.messageId
-        ) ?: return
-        disableLikesForPost(
-            postId,
-            executor,
-            replied.chat.id,
-            postsLikesMessagesTable
-        )
+        try {
+            val postId = PostsTable.findPost(
+                replied.messageId
+            )
+            disableLikesForPost(
+                postId,
+                executor,
+                replied.chat.id,
+                postsLikesMessagesTable
+            )
 
-        commonLogger.info("Rating was disabled")
+            executor.executeUnsafe(
+                DeleteMessage(
+                    message.chat.id,
+                    message.messageId
+                )
+            )
+
+            commonLogger.info("Rating was disabled")
+        } catch (e: NoRowFoundException) {
+            commonLogger.throwing(this::class.simpleName, "disableRating", e)
+
+            executor.executeUnsafe(
+                SendMessage(
+                    message.chat.id,
+                    "Post for message was not found",
+                    replyToMessageId = message.messageId
+                )
+            )
+        }
     }
 }